triboka-agro/backend/requirements.txt
# Triboka BaaS Platform - Backend Dependencies

# Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Blockchain & Web3
web3==6.11.0
eth-account==0.9.0
eth-utils==2.2.1

# Security & Authentication
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0

# API & HTTP
requests==2.31.0
urllib3==2.0.5

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Date & Time
python-dateutil==2.8.2

# Validation
marshmallow==3.20.1
cerberus==1.3.5

# Background Tasks (optional)
celery==5.3.2
redis==5.0.0

# File Upload & Storage
Pillow==10.0.1
boto3==1.29.7  # For AWS S3 if needed

# Monitoring & Logging
structlog==23.1.0

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Production Server
gunicorn==21.2.0
gevent==23.7.0

# IPFS Integration (optional)
ipfshttpclient==0.8.0a2

# Cryptography
cryptography==41.0.4

# Utils
shortuuid==1.0.11
qrcode==7.4.2